# openapi.schema.yaml
openapi: 3.1.0
info:
  title: mehari
  description: Variant effect prediction all in Rust
  contact:
    name: Manuel Holtgrewe
    email: manuel.holtgrewe@bih-charite.de
  license:
    name: MIT
    identifier: MIT
  version: 0.41.2
paths:
...
  /api/v1/versionsInfo:
    get:
      tags:
      - versions
      summary: Query for consequence of a variant.
      operationId: versionsInfo
      responses:
        '200':
          description: Version information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionsInfoResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
components:
  schemas:
...
    CustomError:
      type: object
      required:
      - err
      properties:
        err:
          type: string
    DataVersionEntry:
      type: object
      description: Specification of data version for a given genome build.
      required:
      - assembly
      - version_annotation
      - annotation_name
      properties:
        assembly:
          type: string
          description: Assembly for which the data version is specified.
        version_refseq:
          type:
          - string
          - 'null'
          description: Version of the RefSeq database, if any.
        version_ensembl:
          type:
          - string
          - 'null'
          description: Version of the Ensembl database, if any.
        version_annotation:
          type: string
          description: Version of annotation used.
        annotation_name:
          type: string
          description: Name of the annotation used.
...
    SoftwareVersions:
      type: object
      description: Software version specification.
      required:
      - mehari
      - hgvs_rs
      properties:
        mehari:
          type: string
          description: Version of `mehari`.
        hgvs_rs:
          type: string
          description: Version of the `hgvs` crate.
...
    VersionsInfoResponse:
      type: object
      description: Response of the `/api/v1/version` endpoint.
      required:
      - software
      - data
      properties:
        software:
          $ref: '#/components/schemas/SoftwareVersions'
          description: Software versions specification.
        data:
          type: array
          items:
            $ref: '#/components/schemas/DataVersionEntry'
          description: Data versions specification.